Added first coment for doc, ROS and minor fixes
