+++ control_window_params.hpp
@@ -1,5 +1,5 @@
-#ifndef CONTROLL_WINDOW_H
-#define CONTROLL_WINDOW_H
+#ifndef CONTROLL_WINDOW_HPP
+#define CONTROLL_WINDOW_HPP
 
 #include <string>
 #include <vector>
@@ -23,19 +23,21 @@ struct parameterController {
         void createTrackbars();
 
         void printCurrentConfig();
-        void loadConfig(int configNr);
+        void loadConfig(unsigned long configNr);
 };
 
 struct parameterDescription {
     public:
         const int minValue;
         const int maxValue;
+        
+        int startValue;
+        
         const std::string name;
 
         cv::TrackbarCallback callbackFunction;
 
         int selectedValue = -1;
-        int startValue;
     
         parameterDescription(int minValue, int maxValue, int startValue, std::string name, cv::TrackbarCallback callbackFunction) : minValue(minValue), maxValue(maxValue), startValue(startValue), name(name), callbackFunction(callbackFunction) {
